Split server.py into modules, clean up dead code, add unit tests
